# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для работы с фильмами и актерами в FilmHub. login: admin, password: kek",
        "title": "FilmHub API",
        "contact": {
            "url": "http://www.github.com/avealice"
        }
    },
    "host": "127.0.0.1:8000",
    "basePath": "/",
    "paths": {
        "/api/actor": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает нового актера.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/api/actor"
                ],
                "summary": "Создать актера.",
                "parameters": [
                    {
                        "description": "Данные нового актера",
                        "name": "actor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InputActor"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Актер успешно создан",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Пустой заголовок авторизации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Некорректный метод",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/movie": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новый фильм.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/api/movie"
                ],
                "summary": "Создать фильм",
                "parameters": [
                    {
                        "description": "Данные нового фильма",
                        "name": "movie",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InputMovie"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Фильм создан успешно"
                    },
                    "400": {
                        "description": "Некорректный запрос или данные",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пустой заголовок авторизации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Некорректная роль",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/movie/search": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Выполняет поиск фильмов по указанным критериям (название или актер).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/api/movie/search"
                ],
                "summary": "Поиск фильмов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название фильма для поиска",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Имя актера для поиска",
                        "name": "actor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список фильмов, удовлетворяющих критериям поиска",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.MovieWithActors"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос или данные",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пустой заголовок авторизации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Некорректная роль",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/movie/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получает информацию о фильме по его идентификатору.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/api/movie/{id}"
                ],
                "summary": "Получить информацию о фильме",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор фильма",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о фильме",
                        "schema": {
                            "$ref": "#/definitions/model.MovieWithActors"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос или данные",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пустой заголовок авторизации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Некорректная роль",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Некорректный метод",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет информацию о фильме.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/api/movie/{id}"
                ],
                "summary": "Обновить информацию о фильме",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор фильма",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные о фильме",
                        "name": "movie",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InputMovie"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Фильм обновлен успешно"
                    },
                    "400": {
                        "description": "Некорректный запрос или данные",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пустой заголовок авторизации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Некорректная роль",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Некорректный метод",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет фильм по его идентификатору.",
                "tags": [
                    "/api/movie/{id}"
                ],
                "summary": "Удалить фильм",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор фильма",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Фильм удален успешно"
                    },
                    "400": {
                        "description": "Некорректный запрос или данные",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пустой заголовок авторизации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Некорректная роль",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Некорректный метод",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/movies": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получает список всех фильмов с возможностью сортировки.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/api/movies"
                ],
                "summary": "Получить все фильмы",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Критерий сортировки (title, rating, release_date)",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Порядок сортировки (asc, desc)",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список фильмов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Movie"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос или данные",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пустой заголовок авторизации",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Некорректный метод",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.ErrorResponse": {
            "description": "JSON-структура ответа с сообщением об ошибке.",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.Actor": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "description": "Birth date of the actor",
                    "type": "string"
                },
                "gender": {
                    "description": "Gender of the actor",
                    "type": "string"
                },
                "name": {
                    "description": "Name of the actor",
                    "type": "string"
                }
            }
        },
...
        "model.InputActor": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "description": "Birth date of the actor",
                    "type": "string"
                },
                "gender": {
                    "description": "Gender of the actor",
                    "type": "string"
                },
                "movies": {
                    "description": "Movies associated with the actor",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Movie"
                    }
                },
                "name": {
                    "description": "Name of the actor",
                    "type": "string"
                }
            }
        },
        "model.InputMovie": {
            "type": "object",
            "properties": {
                "actors": {
                    "description": "Actors associated with the movie",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Actor"
                    }
                },
                "description": {
                    "description": "Description of the movie",
                    "type": "string"
                },
                "rating": {
                    "description": "Rating of the movie",
                    "type": "integer"
                },
                "release_date": {
                    "description": "Release date of the movie",
                    "type": "string"
                },
                "title": {
                    "description": "Title of the movie",
                    "type": "string"
                }
            }
        },
        "model.Movie": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "Description of the movie",
                    "type": "string"
                },
                "rating": {
                    "description": "Rating of the movie",
                    "type": "integer"
                },
                "release_date": {
                    "description": "Release date of the movie",
                    "type": "string"
                },
                "title": {
                    "description": "Title of the movie",
                    "type": "string"
                }
            }
        },
        "model.MovieWithActors": {
            "type": "object",
            "properties": {
                "actors": {
                    "description": "Actors associated with the movie",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Actor"
                    }
                },
                "description": {
                    "description": "Description of the movie",
                    "type": "string"
                },
                "id": {
                    "description": "Unique identifier for the movie",
                    "type": "integer"
                },
                "rating": {
                    "description": "Rating of the movie",
                    "type": "integer"
                },
                "release_date": {
                    "description": "Release date of the movie",
                    "type": "string"
                },
                "title": {
                    "description": "Title of the movie",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}